moved SearchActivity to MainActivity
