+++ app/src/main/AndroidManifest.xml
@@ -13,14 +13,6 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
-        <activity android:name=".views.SearchActivity"
-            android:label="@string/page_pal">
-            <intent-filter>
-                <action android:name="android.intent.action.SEARCH" />
-            </intent-filter>
-            <meta-data android:name="android.app.searchable"
-                android:resource="@xml/searchable"/>
-        </activity>
         <activity android:name=".views.EditBookActivity"></activity>
         <activity android:name=".views.BookDetailsActivity"></activity>
         <activity android:name=".views.EditUserActivity" />
@@ -42,10 +34,10 @@
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
+                <action android:name="android.intent.action.SEARCH" />
             </intent-filter>
-            <meta-data
-                android:name="android.app.default_searchable"
-                android:value=".views.SearchActivity" />
+            <meta-data android:name="android.app.searchable"
+                android:resource="@xml/searchable"/>
         </activity>
     </application>
