+++ app/src/main/java/ca/team21/pagepal/views/MainActivity.java
@@ -23,8 +23,10 @@ import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
 
+import java.util.ArrayList;
 import java.util.Date;
 
 import ca.team21.pagepal.R;
@@ -40,7 +42,8 @@ public class MainActivity extends AppCompatActivity
         BorrowingFragment.OnBorrowingInteractionListener,
         BookFragment.OnListFragmentInteractionListener,
         NotificationsFragment.OnNotificationsInteractionListener,
-        ProfileFragment.OnProfileInteractionListener {
+        ProfileFragment.OnProfileInteractionListener,
+        SearchedResultsFragment.OnSearchFragmentInteractionListener {
 
     private static final String TAG = "MainActivity";
     private static final int EDIT_USER = 9;
@@ -102,6 +105,11 @@ public class MainActivity extends AppCompatActivity
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
+        if (Intent.ACTION_SEARCH.equals(getIntent().getAction())) {
+            final String keyword = getIntent().getStringExtra(SearchManager.QUERY);
+            queryBooks(keyword);
+        }
+
         // Wait for user to authenticate or timeout
         long waitTime = new Date().getTime() + 2 * 1000;
         while ( authUser == null && new Date().getTime() < waitTime ) {
@@ -129,6 +137,55 @@ public class MainActivity extends AppCompatActivity
 
         loadFragment(HomeFragment.newInstance());
 
+
+
+    }
+    /**
+     * Queries the firebase realtime database, filters the results, maps them to Books and adds the Books to a list.
+     *
+     * @param query a string representation of the user's search query
+     */
+    public void queryBooks(final String query) {
+
+        final ArrayList<Book> bookList = new ArrayList<Book>();
+        final String[] keyWords = query.split("\\s+");
+        // Query Firebase
+        Query bookQuery = FirebaseDatabase.getInstance().getReference().child("books");
+
+        bookQuery.addListenerForSingleValueEvent(new ValueEventListener() {
+
+            @Override
+            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                if (dataSnapshot.exists()) {
+                    for (DataSnapshot users: dataSnapshot.getChildren()) { // for each user
+                        for (DataSnapshot data: users.getChildren()) { // for each of their books
+                            String status = data.child("status").getValue(String.class);
+                            // Filter by Status
+                            if (status != null && ( status.equals("Available") || status.equals("Requested"))) {
+
+                                Book book = data.getValue(Book.class);
+                                for (String keyWord : keyWords) {
+
+                                    if ((book.getAuthor().toUpperCase()).contains(keyWord.toUpperCase()) ||
+                                            (book.getTitle().toUpperCase()).contains(keyWord.toUpperCase()) ||
+                                            (book.getDescription().toUpperCase()).contains(keyWord.toUpperCase())) {
+                                        bookList.add(book);
+                                    }
+                                }
+                            }
+                        }
+                    }
+                    loadFragment(SearchedResultsFragment.newInstance(bookList));
+
+                }
+
+            }
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError){
+                Log.w(TAG, "queryBooks:failure", databaseError.toException());
+            }
+
+        });
 
 
     }
@@ -236,4 +293,12 @@ public class MainActivity extends AppCompatActivity
     }
 
 
+    @Override
+    public void viewBookInteraction(Book book) {
+        Intent intent = new Intent(this, BookDetailsActivity.class);
+        intent.putExtra(MainActivity.BOOK_EXTRA, book);
+        User requester = User.getInstance();
+        intent.putExtra(MainActivity.USER_EXTRA, requester);
+        startActivity(intent);
+    }
 }
